[lldb] Allow for the evaluation of self in a generic context
There are situations where the compiler may not know all the types
involved in evaluating an expression. For example, lldb might be stopped
at a point where self is parameterized by some private type which lldb
can't reconstruct the AST type. This patch allows for evaluating
expressions with such types by not attempting to bind the generic types,
this behavior is controlled by a new option.

The approach taken by this patch is setting up a trampoline function that
calls the method with the user's code, and a sink function, which the
entry point calls, and is later used to redirect the call to the
trampoline function (at the LLVM-IR level).

rdar://91417802